game/requirements.txt
pygame~=2.5.2
pillow~=10.2.0
# python 3.8